feat(tests): add new shader test files for debugging and validation
- Introduced multiple new test files to enhance shader testing, including `fa3-inout.glsl`, `fa3-mat4-trace.glsl`, and additional files for edge cases.
- Updated existing test files to improve validation of shader functionality and ensure comprehensive coverage.
- Adjusted output types in tests to align with expected results, enhancing the robustness of the testing framework.

These additions improve the clarity and effectiveness of shader testing, facilitating better validation of shader behavior.

+++ lp-shader/lps-filetests/filetests/debug/fa3-out-not-read.glsl
@@ -6,7 +6,7 @@ void set_components(out ivec3 result) {
     result.z = 3;
 }
 
-float test_edge_out_modify_components() {
+int test_edge_out_modify_components() {
     // Can modify individual components without reading whole vector
     ivec3 v;
     set_components(v);
